refactor(experimentalist): move Layer A entities out of components/evaluator/
`Dataset`, `Task`, `TrialResult`, `EvaluationResult`, `DatasetRef`, `ResourceRef`
and the metric/dependency models lived in
`experimentalist/components/evaluator/models.py` — Layer A filed inside a Layer B
component. They are the contract every plugin speaks, and the evaluator is one
consumer among several, so they now sit in `entities.py` beside `ExperimentRun`
and `Candidate`.

Pure relocation: no field, validator or behaviour changes. 43 files repointed,
no re-export shim.

This also reclassifies the Eval Author boundary ratchet, which previously counted
imports without distinguishing what they were for. It now pins two lists:

- `_SHARED_LAYER_A` — the entity contract. Permanent by design. Duplicating it
  would fork the contract and break Studio comparability, which is the one thing
  the design cannot trade away.
- `_BORROWED_BEHAVIOUR` — Harbor, tools, trace analysis, the backend factory.
  Still debt, still shrink-only, still "duplicate rather than add a row".

Five of the twelve original rows were Layer A and collapse into one shared entry;
the reverse assertion now guards only the debt list, so removing a borrowed module
still fails loudly while sharing entities does not.
