+++ plugins/nemo-eval-author/src/nemo_eval_author_plugin/eval_author/models.py
@@ -3,7 +3,7 @@
 
 """Shared models for the top-level Eval Author."""
 
-from nemo_experimentalist_plugin.experimentalist.components.evaluator import Dataset
+from nemo_experimentalist_plugin.entities import Dataset
 from pydantic import BaseModel, ConfigDict, Field
