+++ plugins/nemo-eval-author/tests/test_eval_author_materialization.py
@@ -12,7 +12,7 @@ from pathlib import Path
 import pytest
 from nemo_eval_author_plugin.eval_author import materialization as materialization_module
 from nemo_eval_author_plugin.eval_author.materialization import InsightSuite
-from nemo_experimentalist_plugin.experimentalist.components.evaluator import Task
+from nemo_experimentalist_plugin.entities import Task
 from nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor import HarborDataset
