+++ plugins/nemo-eval-author/tests/test_eval_author_repair_e2e.py
@@ -12,10 +12,7 @@ import pytest
 from nemo_eval_author_plugin.eval_author.agent import EvalAuthor
 from nemo_eval_author_plugin.eval_author.models import EvalAuthorConfig
 from nemo_eval_author_plugin.model_config import get_fast_model
-from nemo_experimentalist_plugin.experimentalist.components.evaluator import (
-    DatasetValidationError,
-    local_path_from_uri,
-)
+from nemo_experimentalist_plugin.entities import DatasetValidationError, local_path_from_uri
 from nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor import (
     HarborDataset,
     HarborEvaluator,
