+++ plugins/nemo-eval-author/tests/test_eval_author_run.py
@@ -8,7 +8,7 @@ from typing import Any, Literal
 import pytest
 from nemo_eval_author_plugin.eval_author import run as eval_author_run
 from nemo_eval_author_plugin.eval_author.models import EvalAuthorConfig, EvalAuthorResult
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import Dataset, DatasetRef, Task
+from nemo_experimentalist_plugin.entities import Dataset, DatasetRef, Task
 from nemo_insights_plugin.entities import Insight
