+++ plugins/nemo-eval-author/tests/test_plugin_boundary.py
@@ -3,14 +3,17 @@
 
 """Ratchet on Eval Author's remaining dependency on Experimentalist.
 
-Eval Author is meant to end up standalone, with Experimentalist depending on it and not the
-other way around. Until then this test pins the modules Eval Author still reaches for, so
-the coupling can only shrink. Prefer duplicating a helper over adding a row here.
+Two lists, because "imports Experimentalist" is not one thing.
 
-Emptying the allowlist is what unblocks the remaining TODO(eval-author-standalone)
-cleanup: the ``EXPERIMENTALIST_*`` credential fallback and the
-``bridge_author_env_to_experimentalist`` call in ``EvalAuthor.__init__`` exist only to
-serve the imports listed below.
+``_SHARED_LAYER_A`` is the entity contract — ``Dataset``, ``Task``, ``TrialResult``,
+``DatasetRef`` and friends. Every plugin is *supposed* to speak these; duplicating them
+would fork the contract and break comparability in Studio. This list does not shrink.
+
+``_BORROWED_BEHAVIOUR`` is debt: Harbor, tools, trace analysis, the backend factory. Eval
+Author is meant to end up standalone, so this list can only shrink, and duplicating a
+helper beats adding a row. Emptying it is what unblocks the remaining
+TODO(eval-author-standalone) cleanup — the ``EXPERIMENTALIST_*`` credential fallback and
+the ``bridge_author_env_to_experimentalist`` call in ``EvalAuthor.__init__``.
 """
 
 import ast
@@ -18,32 +21,31 @@ from pathlib import Path
 
 _SRC = Path(__file__).resolve().parents[1] / "src" / "nemo_eval_author_plugin"
 
-# Each entry is the module named in the import statement, so `from ...components import
-# cache` is recorded as `...components`. Each one is debt: deleting a row is the goal, and
-# adding a row needs a deliberate argument for why duplicating the helper is worse. What
-# Eval Author still borrows, by row:
+# Layer A: the shared entity contract. Permanent by design, not debt.
+_SHARED_LAYER_A = {
+    "nemo_experimentalist_plugin.entities",
+}
+
+# Behaviour Eval Author still borrows. Each row is debt; deleting one is the goal, and
+# adding one needs a deliberate argument for why duplicating the helper is worse.
 #
-#   client              -> make_client, the platform client factory
-#   ...components       -> the cache module, for run artifacts
-#   ...dataset_staging  -> stage_task_template
-#   ...evaluator        -> Dataset / Task / TrialResult / DatasetValidationError
-#   ...evaluator.base   -> EvaluatorType
-#   ...evaluator.factory-> DatasetFactory
-#   ...evaluator.harbor -> HarborDataset
-#   ...evaluator.models -> DatasetRef, ResourceRef, Task, local_path_from_uri
-#   ...tools            -> GuardedShellTools
-#   ...trace_analyzer   -> TraceAnalyzer, TraceAnalyzerConfig, Diagnostic
-#   ...trace_explorer   -> TraceExplorer and its view models
+#   client                  -> make_client, the platform client factory
+#   ...components           -> the cache module, for run artifacts
+#   ...dataset_staging      -> stage_task_template
+#   ...evaluator.base       -> EvaluatorType
+#   ...evaluator.factory    -> DatasetFactory
+#   ...evaluator.harbor     -> HarborDataset
+#   ...tools                -> GuardedShellTools
+#   ...trace_analyzer       -> TraceAnalyzer, TraceAnalyzerConfig, Diagnostic
+#   ...trace_explorer       -> TraceExplorer and its view models
 #   experimentalist_backend -> make_experimentalist_backend
-_ALLOWED_EXPERIMENTALIST_IMPORTS = {
+_BORROWED_BEHAVIOUR = {
     "nemo_experimentalist_plugin.client",
     "nemo_experimentalist_plugin.experimentalist.components",
     "nemo_experimentalist_plugin.experimentalist.components.dataset_staging",
-    "nemo_experimentalist_plugin.experimentalist.components.evaluator",
     "nemo_experimentalist_plugin.experimentalist.components.evaluator.base",
     "nemo_experimentalist_plugin.experimentalist.components.evaluator.factory",
     "nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor",
-    "nemo_experimentalist_plugin.experimentalist.components.evaluator.models",
     "nemo_experimentalist_plugin.experimentalist.components.tools",
     "nemo_experimentalist_plugin.experimentalist.components.trace_analyzer",
     "nemo_experimentalist_plugin.experimentalist.components.trace_explorer",
@@ -72,21 +74,22 @@ def _imported_experimentalist_modules() -> dict[str, set[str]]:
 
 def test_experimentalist_imports_only_shrink() -> None:
     found = _imported_experimentalist_modules()
+    allowed = _SHARED_LAYER_A | _BORROWED_BEHAVIOUR
 
-    added = set(found) - _ALLOWED_EXPERIMENTALIST_IMPORTS
+    added = set(found) - allowed
     assert not added, (
         "New Experimentalist imports in the Eval Author plugin: "
         + ", ".join(f"{name} (in {', '.join(sorted(found[name]))})" for name in sorted(added))
         + ". Eval Author is moving toward standalone, so duplicate the helper instead of "
-        "importing it. If the import is genuinely unavoidable, add it to "
-        "_ALLOWED_EXPERIMENTALIST_IMPORTS with a note explaining why."
+        "importing it. If it is Layer A, add it to _SHARED_LAYER_A; if the borrow is "
+        "genuinely unavoidable, add it to _BORROWED_BEHAVIOUR with a note explaining why."
     )
 
-    removed = _ALLOWED_EXPERIMENTALIST_IMPORTS - set(found)
+    removed = _BORROWED_BEHAVIOUR - set(found)
     assert not removed, (
-        "These Experimentalist imports are gone, which is the point: "
+        "These borrowed Experimentalist modules are gone, which is the point: "
         + ", ".join(sorted(removed))
-        + ". Drop them from _ALLOWED_EXPERIMENTALIST_IMPORTS so the list keeps meaning something."
+        + ". Drop them from _BORROWED_BEHAVIOUR so the list keeps meaning something."
     )
