+++ plugins/nemo-experimentalist/src/nemo_experimentalist_plugin/experimentalist/components/evaluator/__init__.py
@@ -1,12 +1,7 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.base import (
-    Evaluator,
-    EvaluatorConfig,
-    EvaluatorType,
-)
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import (
+from nemo_experimentalist_plugin.entities import (
     CommandSpec,
     Dataset,
     DatasetRef,
@@ -22,6 +17,11 @@ from nemo_experimentalist_plugin.experimentalist.components.evaluator.models imp
     TrialResult,
     TrialStatus,
     local_path_from_uri,
+)
+from nemo_experimentalist_plugin.experimentalist.components.evaluator.base import (
+    Evaluator,
+    EvaluatorConfig,
+    EvaluatorType,
 )
 
 __all__ = [
