+++ plugins/nemo-experimentalist/src/nemo_experimentalist_plugin/experimentalist/components/insight_promotion.py
@@ -10,14 +10,14 @@ from collections.abc import Sequence
 from dataclasses import dataclass, replace
 from pathlib import Path
 
-from nemo_experimentalist_plugin.entities import Candidate
-from nemo_experimentalist_plugin.experimentalist.components.evaluator import (
+from nemo_experimentalist_plugin.entities import (
+    Candidate,
     Dataset,
     EvaluationResult,
     Task,
     TrialResult,
+    local_path_from_uri,
 )
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import local_path_from_uri
 
 _GENERIC_METRIC_NAMES = frozenset({"reward", "score"})
 _MAX_REPEAT_SPREAD = 0.1
