+++ plugins/nemo-experimentalist/src/nemo_experimentalist_plugin/experimentalist/components/rationalizer.py
@@ -9,8 +9,7 @@ from collections import Counter, defaultdict  # noqa: F401
 from pathlib import Path
 from typing import Any
 
-from nemo_experimentalist_plugin.experimentalist.components.evaluator import Task
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import DependencyRuntime
+from nemo_experimentalist_plugin.entities import DependencyRuntime, Task
 from nooa import Agent, CodeActStrategy, strategy
 from nooa.agentdoc import doc, spec
 from nooa.agents import TokenBudgetSummarizer
