+++ plugins/nemo-experimentalist/src/nemo_experimentalist_plugin/experimentalist/components/terminator.py
@@ -17,6 +17,7 @@ from typing import Any
 # Imported from `resolve` rather than `.loop`, which merely re-exports it: `loop` imports
 # this module, so going through it would be circular.
 from nemo_experimentalist_plugin.config import EvolutionaryOptimizerConfig
+from nemo_experimentalist_plugin.experimentalist.components.models import EvolutionTree, pareto_front
 from nemo_experimentalist_plugin.skills import skills_dir
 from nooa import Agent, CodeActStrategy, TextSkill, hidden, strategy
 from nooa.agentdoc import doc
@@ -24,7 +25,6 @@ from nooa.config import CodeActConfig
 from pydantic import BaseModel
 
 from .model_config import get_fast_model
-from .models import EvolutionTree, pareto_front
 
 logger = logging.getLogger(__name__)
