+++ plugins/nemo-experimentalist/src/nemo_experimentalist_plugin/experimentalist/components/trace_analyzer.py
@@ -11,9 +11,8 @@ from pathlib import Path
 from typing import Any, Literal
 from urllib.parse import unquote, urlparse
 
+from nemo_experimentalist_plugin.entities import DependencyRuntime, MetricResult, Task, TrialResult
 from nemo_experimentalist_plugin.experimentalist.components import cache
-from nemo_experimentalist_plugin.experimentalist.components.evaluator import MetricResult, Task, TrialResult
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import DependencyRuntime
 from nemo_experimentalist_plugin.experimentalist.components.model_config import get_fast_model, get_smart_model
 from nemo_experimentalist_plugin.experimentalist.components.tools import GuardedShellTools, WorkspaceTool
 from nemo_insights_plugin.entities import Insight
