+++ plugins/nemo-experimentalist/src/nemo_experimentalist_plugin/experimentalist/run.py
@@ -8,8 +8,8 @@ import logging
 from pathlib import Path
 from typing import Protocol, TextIO, cast
 
+from nemo_experimentalist_plugin.entities import DatasetRef
 from nemo_experimentalist_plugin.experimentalist.agent import build_experimentalist_agent
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import DatasetRef
 from nemo_experimentalist_plugin.experimentalist.components.loop import EvolutionaryOptimizerConfig
 from nemo_experimentalist_plugin.experimentalist.deps import ExperimentalistDeps
 from nemo_experimentalist_plugin.experimentalist.experimentalist_backend import (
