+++ plugins/nemo-experimentalist/src/nemo_experimentalist_plugin/resolve.py
@@ -25,7 +25,7 @@ import yaml
 from nemo_experimentalist_plugin.config import (
     EvolutionaryOptimizerConfig,
 )
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import DatasetRef
+from nemo_experimentalist_plugin.entities import DatasetRef
 from nemo_experimentalist_plugin.experimentalist.components.repository import looks_like_git
 from nemo_experimentalist_plugin.profile import AgentProfile
 from nemo_insights_plugin.contracts.insights import InsightsFileError, load_insights_document
