+++ plugins/nemo-experimentalist/tests/experimentalist/test_loop_reporting.py
@@ -7,16 +7,16 @@ from types import SimpleNamespace
 from unittest.mock import AsyncMock
 
 import pytest
-from nemo_experimentalist_plugin.entities import Candidate
-from nemo_experimentalist_plugin.experimentalist.components.evaluator import (
+from nemo_experimentalist_plugin.entities import (
+    Candidate,
     Dataset,
     EvaluationResult,
     MetricResult,
+    ResourceRef,
     Task,
     TrialResult,
-    TrialStatus,
 )
-from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import ResourceRef
+from nemo_experimentalist_plugin.experimentalist.components.evaluator import TrialStatus
 from nemo_experimentalist_plugin.experimentalist.components.insight_promotion import (
     _task_evidence,
     insight_suite_provenance,
