+++ plugins/nemo-experimentalist/tests/test_otlp.py
@@ -14,7 +14,7 @@ from datetime import datetime, timezone
 from pathlib import Path
 
 import pytest
-from nemo_experimentalist_plugin.experimentalist.components.evaluator import ResourceRef
+from nemo_experimentalist_plugin.entities import ResourceRef
 from nemo_experimentalist_plugin.experimentalist.otlp import (
     jsonl_to_protobuf,
     read_trace_id,
